requirements/base.txt
Django==2.0.5
psycopg2-binary==2.7.4
docker==3.5.1

django-import-export==1.0.0
django-celery-beat==1.1.1
django-celery-results==1.0.1

djangorestframework==3.8.2
djangorestframework-datatables==0.3.0
django-bootstrap4==0.0.6

python-bittrex==0.3.0
celery==4.1.1
redis==2.10.6
uuid==1.30
python-dotenv==0.9.1
flake8==3.6.0
